Refactored component names to be explicit, updated mobile menu
