+++ public/index.html
@@ -5,7 +5,10 @@
         <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <meta name="theme-color" content="#000000" />
-        <meta name="description" content="Benjamin (Ben) Weare's Dashboard" />
+        <meta
+            name="description"
+            content="This page was created by Benjamin (Ben) Weare; it serves as a dashboard for collecting news of interest. Information about Ben can also be found here."
+        />
 
         <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
