+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-    "short_name": "Ben's web page",
-    "name": "Ben's web page",
+    "short_name": "Benjamin (Ben) Weare's web page",
+    "name": "Benjamin (Ben) Weare's web page",
     "start_url": ".",
     "display": "standalone",
     "theme_color": "#000000",
