+++ src/common/hooks/externalClickHandler.ts
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import { useEffect } from "react";
 
 /**
  * This function tracks the focus of an element
