+++ src/common/hooks/swipeHandler.ts
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 
 /**
  * This function tracks the swipe movement of an element
@@ -11,8 +11,7 @@ export const SwipeHandler = (
 ) => {
     const [touchStart, setTouchStart] = useState<number | null>();
 
-    // TODO: Check for Y changes so page doesn't scroll while swiping
-
+    // TODO -> Check for Y changes so page doesn't scroll while swiping
     useEffect(() => {
         const handleSwipeStart = ({ target, targetTouches }: TouchEvent) => {
             if (ref && ref.current && ref.current.contains(target as Node)) {
