+++ src/common/types.ts
@@ -1,5 +1,3 @@
-import React from "react";
-
 /* TYPES FOR CLIENT */
 export interface Icons {
     Home: React.FunctionComponent;
