+++ src/components/home/pages/info/info.test.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Icon } from "src/components/";
 import { act, render, screen } from "@testing-library/react";
 import Component from "./info";
