+++ src/components/index.ts
@@ -1,13 +1,14 @@
+// Pages:
 export { default as Info } from "./home/pages/info/info";
 export { default as Dashboard } from "./home/pages/dashboard/dashboard";
 export { default as Routes } from "./home/home";
-export { default as NewsCarousel } from "./shared/news-reel/news-reel";
-export { default as Card } from "./shared/card/card";
-export { default as Loader } from "./shared/loader/loader";
-export { default as Error } from "./shared/error/error";
-export * as Icon from "./shared/icons/icons";
-export {
-    NavigationBar,
-    LeftNavigationBar,
-    RightNavigationBar,
-} from "./shared/navigation/navigation";
+
+// Components:
+export { default as NewsCarousel } from "./shared/news-reel/news-reel-component";
+export { default as Card } from "./shared/card/card-component";
+export { default as Loader } from "./shared/loader/loader-component";
+export { default as Error } from "./shared/error/error-component";
+export { default as NavigationBar } from "./shared/navigation-top/navigation-top-component";
+export { default as LeftNavigationBar } from "./shared/navigation-left/navigation-left-component";
+export { default as RightNavigationBar } from "./shared/navigation-right/navigation-right-component";
+export * as Icon from "./shared/icons/icon-component";
