+++ src/components/shared/card/card-component.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { NewsArticle, NewsCard } from "src/common/types";
 import { sleep } from "src/common/utils";
 import { Error, Loader } from "src/components/";
@@ -26,15 +26,15 @@ const Card = ({ Endpoint, SiteName }: NewsCard) => {
 
     return (
         <div className="px-2 md:px-6 my-3 w-full">
-            <div className="text-left flex flex-col w-full items-center justify-center md:p-4 md:border border-slate-300 dark:border-zinc-600/20 rounded-xl">
+            <div className="text-left flex flex-col w-full items-center justify-center md:p-4 md:border border-slate-300 dark:border-zinc-600/20 rounded">
                 {loaded === true &&
                     article &&
                     (!article.url.includes("youtube.com") ? (
                         <div
-                            className={`animate__animated animate__fadeIn animate__faster w-full rounded-xl flex-col xl:flex-row bg-white dark:bg-zinc-900/80 shadow-md`}
+                            className={`animate__animated animate__fadeIn animate__faster w-full rounded flex-col xl:flex-row bg-white dark:bg-zinc-900/70 shadow-md`}
                         >
                             <div
-                                className="rounded-t-xl w-full h-52 shadow-sm bg-cover"
+                                className="rounded-t w-full h-52 shadow-sm bg-cover"
                                 style={{
                                     backgroundImage: `url(${article.url})`,
                                 }}
@@ -51,7 +51,7 @@ const Card = ({ Endpoint, SiteName }: NewsCard) => {
                                         onClick={() =>
                                             window.open(article.url, "_blank")
                                         }
-                                        className="transition-all duration-100 text-center p-2 rounded-md text-white w-1/2 bg-gradient-to-r from-blue-700 to-blue-500 hover:shadow-md hover:from-blue-800 hover:to-blue-600"
+                                        className="transition-all duration-100 text-center p-2 rounded text-white w-1/2 bg-gradient-to-r from-blue-700 to-blue-500 hover:shadow-md hover:from-blue-800 hover:to-blue-600"
                                     >
                                         View Full-Size
                                     </button>
@@ -72,7 +72,7 @@ const Card = ({ Endpoint, SiteName }: NewsCard) => {
                             href={article.url}
                             rel="noreferrer"
                             target="_blank"
-                            className={`animate__animated animate__fadeIn animate__faster w-full rounded-xl flex-col xl:flex-row bg-white dark:bg-zinc-700 shadow-md`}
+                            className={`animate__animated animate__fadeIn animate__faster w-full rounded flex-col xl:flex-row bg-white dark:bg-zinc-700 shadow-md`}
                         >
                             <div className="p-2">
                                 <iframe
