+++ src/components/shared/loader/loader-component.tsx
